[#3] refactor: project type을 root type(@types)에서 type 디렉토리로 변경
